chore: Update BASE_URL references and refactor code for consistency
