+++ frontend/src/App.jsx
@@ -21,19 +21,15 @@ const ProtectedLayout = () => {
   );
 };
 
-// export const BASE_URL =
-//   "https://fictional-space-sniffle-94pgrxqq9qwfxg54-3000.app.github.dev";
-// export const BASE_URL =
 
-//   "https://fictional-space-sniffle-94pgrxqq9qwfxg54-3000.app.github.dev";
-// export const BASE_URL = "http://127.0.0.1:3000";
-export const BASE_URL = "";
+
+
 const App = () => {
   return (
 <Router basename="/">
   <Layout>
     <Routes initialRoute="home">
-      <Route path="/login" element={<Login />} />
+
       <Route path="home" element={<Home />} />
       <Route path="studentTable" element={<StudentTable />} />
     </Routes>
