+++ frontend/src/lib/db.ts
@@ -1,0 +1,4 @@
+
+const BASE_URL = "http://127.0.0.1:3000";
+
+export default BASE_URL;
