+++ frontend/src/pages/Login.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import axios from "axios";
-import { BASE_URL } from "../App";
+import BASE_URL from "../lib/db"
 import { useNavigate } from "react-router-dom";
 const Login = () => {
   const [loginData, setLoginData] = useState({
@@ -29,7 +29,9 @@ const Login = () => {
 
   return (
     <div className="flex items-center justify-center min-h-screen bg-gray-100">
-      <form className="bg-white p-6 rounded-lg shadow-md w-full max-w-sm">
+      <form className="bg-white p-6 rounded-lg shadow-md w-full max-w-sm"
+      onSubmit={handleSubmit}
+      >
         <h2 className="text-2xl font-bold mb-4">Login</h2>
 
         <input
@@ -53,7 +55,8 @@ const Login = () => {
         <button
           type="submit"
           className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-          onClick={handleSubmit}
+
+
         >
           Login
         </button>
